+++ CHANGELOG.md
@@ -92,6 +92,7 @@
 - Binance parser: error unrecognised operations.
 - Conversion tool: updated mappings for RECAP output format.
 - Bitstamp parser: added "Staking reward" transaction type.
+- Crypto.com parser: added additional transaction kinds. ([#285](https://github.com/BittyTax/BittyTax/issues/285))
 
 ## Version [0.5.0] Beta (2021-11-11)
 Important:-